refactor(initializer): separate component code into non index file
